Move decorator to a decorators.ts file
