+++ packages/web-components/@groupby/elements-base/src/index.ts
@@ -1,2 +1,3 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as Base, dataInitializer } from './base';
+export { default as Base } from './base';
+export { dataInitializer } from './decorators';
